# swagger-ui/api-docs.yaml
openapi: 3.0.3
info:
  title: Todo App REST API
  description: |
    A comprehensive REST API for a Todo application built with Express.js and JSON Server.
    
    ## Features
    - User registration and authentication
    - Todo CRUD operations
    - Category management
    - User-Todo relationships
    - JWT-based authentication
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@todoapp.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api-proxy
    description: Proxied API Server (Recommended for Codespaces)
  - url: http://localhost:5000
    description: Local Development Express Server
  - url: http://localhost:3001
    description: Local Development JSON Server
  - url: http://localhost:3001
    description: JSON Server (Database)
...
paths:
...
  /api/users/profile:
    get:
      tags:
        - Users
      summary: Get user profile
      description: Get the profile of the authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'

    put:
      tags:
        - Users
      summary: Update user profile
      description: Update the profile of the authenticated user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                  minLength: 1
                  maxLength: 50
                  example: "John"
                lastName:
                  type: string
                  minLength: 1
                  maxLength: 50
                  example: "Doe"
                email:
                  type: string
                  format: email
                  example: "john.doe@example.com"
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Profile updated successfully"
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
  /api/users/preferences:
    patch:
      tags:
        - Users
      summary: Update user preferences
      description: Update user preferences like theme, language, notifications
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                preferences:
                  $ref: '#/components/schemas/UserPreferences'
      responses:
        '200':
          description: Preferences updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Preferences updated successfully"
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
  /api/users/stats:
    get:
      tags:
        - Users
      summary: Get user statistics
      description: Get statistics about user's todos
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User statistics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      stats:
                        $ref: '#/components/schemas/UserStats'
  /api/users/deactivate:
    post:
      tags:
        - Users
      summary: Deactivate user account
      description: Deactivate the user account
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Account deactivated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/users/account:
    delete:
      tags:
        - Users
      summary: Delete user account
      description: Permanently delete user account and all associated data
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Account deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # Todo Endpoints
  /api/todos:
    get:
      tags:
        - Todos
      summary: Get todos
      description: Get todos for the authenticated user with filtering and pagination
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filter by todo status
          schema:
            type: string
            enum: [pending, in-progress, completed]
        - name: priority
          in: query
          description: Filter by todo priority
          schema:
            type: string
            enum: [low, medium, high, urgent]
        - name: category
          in: query
          description: Filter by category name
          schema:
            type: string
        - name: search
          in: query
          description: Search in title, description, and tags
          schema:
            type: string
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: sortBy
          in: query
          description: Field to sort by
          schema:
            type: string
            default: createdAt
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Todos retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      todos:
                        type: array
                        items:
                          $ref: '#/components/schemas/Todo'
                      pagination:
                        $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Todos
      summary: Create todo
      description: Create a new todo for the authenticated user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoInput'
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Todo created successfully"
                  data:
                    type: object
                    properties:
                      todo:
                        $ref: '#/components/schemas/Todo'
...
  /todos:
    get:
      tags:
        - Relations
      summary: Get all todos (JSON Server)
      description: Direct access to JSON Server todos endpoint
      parameters:
        - name: userId
          in: query
          description: Filter todos by user ID
          schema:
            type: string
      responses:
        '200':
          description: Todos retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: "1"
        username:
          type: string
          example: "johndoe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
        firstName:
          type: string
          example: "John"
        lastName:
          type: string
          example: "Doe"
        avatar:
          type: string
          example: ""
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        loginAttempts:
          type: integer
          example: 0
        accountLockedUntil:
          type: string
          format: date-time
          nullable: true
        isActive:
          type: boolean
          example: true
        preferences:
          $ref: '#/components/schemas/UserPreferences'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserPreferences:
      type: object
      properties:
        theme:
          type: string
          enum: [light, dark, system]
          example: "light"
        language:
          type: string
          example: "en"
        notifications:
          type: object
          properties:
            browser:
              type: boolean
              example: true
            inApp:
              type: boolean
              example: true
        timezone:
          type: string
          example: "UTC"
    UserStats:
      type: object
      properties:
        totalTodos:
          type: integer
          example: 25
        completedTodos:
          type: integer
          example: 15
        pendingTodos:
          type: integer
          example: 8
        inProgressTodos:
          type: integer
          example: 2
        overdueTodos:
          type: integer
          example: 3
        completionRate:
          type: integer
          example: 60
        categoryCounts:
          type: object
          example:
            Work: 10
            Personal: 8
            Shopping: 7
        priorityCounts:
          type: object
          example:
            high: 5
            medium: 12
            low: 8
    Todo:
      type: object
      properties:
        id:
          type: string
          example: "1"
        userId:
          type: string
          example: "1"
        title:
          type: string
          example: "Complete project documentation"
        description:
          type: string
          example: "Write comprehensive API documentation for the todo app"
        status:
          type: string
          enum: [pending, in-progress, completed]
          example: "in-progress"
        priority:
          type: string
          enum: [low, medium, high, urgent]
          example: "high"
        category:
          type: string
          example: "Work"
        tags:
          type: array
          items:
            type: string
          example: ["documentation", "api", "urgent"]
        dueDate:
          type: string
          format: date-time
          nullable: true
        reminderDate:
          type: string
          format: date-time
          nullable: true
        estimatedTime:
          type: integer
          nullable: true
          description: "Estimated time in minutes"
        actualTime:
          type: integer
          nullable: true
          description: "Actual time spent in minutes"
        isRecurring:
          type: boolean
          example: false
        recurringPattern:
          type: object
          nullable: true
        parentTodoId:
          type: string
          nullable: true
        attachments:
          type: array
          items:
            type: object
        comments:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              text:
                type: string
              createdAt:
                type: string
                format: date-time
        dependencies:
          type: object
          properties:
            blockedBy:
              type: array
              items:
                type: string
            blocking:
              type: array
              items:
                type: string
        completedAt:
          type: string
          format: date-time
          nullable: true
        isArchived:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TodoInput:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: "Complete project documentation"
        description:
          type: string
          maxLength: 1000
          example: "Write comprehensive API documentation for the todo app"
        status:
          type: string
          enum: [pending, in-progress, completed]
          example: "pending"
        priority:
          type: string
          enum: [low, medium, high, urgent]
          example: "medium"
        category:
          type: string
          maxLength: 50
          example: "Work"
        tags:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 30
        dueDate:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        currentPage:
          type: integer
          example: 1
        totalPages:
          type: integer
          example: 5
        totalItems:
          type: integer
          example: 47
        limit:
          type: integer
          example: 10
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
...